Exception handling - delete

+++ src/main/java/com/kamishhayha/coursek/services/UserService.java
@@ -4,48 +4,55 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.kamishhayha.coursek.entities.User;
 import com.kamishhayha.coursek.repositories.UserRepository;
+import com.kamishhayha.coursek.services.exceptions.DatabaseException;
 import com.kamishhayha.coursek.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class UserService {
-	
+
 	@Autowired
-	private UserRepository repository; 
-	
+	private UserRepository repository;
+
 	public List<User> findAll() {
 		return repository.findAll();
 	}
-	
+
 	public User findByid(Long id) {
 		Optional<User> user = repository.findById(id);
 		return user.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User user) {
 		return repository.save(user);
 	}
-	
+
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if (!repository.existsById(id)) {
+				throw new ResourceNotFoundException(id);
+			}
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
-	
+
 	public User update(Long id, User user) {
 		User entity = repository.getReferenceById(id);
 		updateData(entity, user);
-		return repository.save(entity); 
+		return repository.save(entity);
 	}
 
 	private void updateData(User entity, User user) {
 		entity.setName(user.getName());
 		entity.setEmail(user.getEmail());
 		entity.setPhone(user.getPhone());
-		
+
 	}
-	
-	
-	
+
 }
